requirements.txt
absl-py==1.1.0
asgiref==3.5.2
astunparse==1.6.3
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
djangorestframework==3.13.1
flatbuffers==1.12
gast==0.4.0
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.12.0
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.1
Markdown==3.4
numpy==1.23.1
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.10
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.1
